Restructured the app into a package rather than a single module
