+++ README.md
@@ -1,6 +1,6 @@
 # collectorSite
 CodeHub project
 
-To run, type "python backEnd\app.py" in the terminal from the root directory
+To run, type "python backEnd\run.py" in the terminal from the root directory
 
 Add other important info here as needed
