+++ backEnd/collectorSite/routes.py
@@ -1,0 +1,27 @@
+from flask import render_template
+from collectorSite import app
+# from collectorSite.models import Product
+
+# simulates data in the database for now
+products = [
+    {
+        'name': 'Keyboard',
+        'price': 100,
+        'description': 'A funny description'
+    },
+    {
+        'name': 'Shoes',
+        'price': 200,
+        'description': 'Test'
+    }
+]
+
+
+@app.route('/')
+def index():
+    return render_template('index.html')
+
+
+@app.route('/products')
+def displayProducts():
+    return render_template('products.html', products=products)
