+++ backEnd/run.py
@@ -1,0 +1,5 @@
+# imports from the __init__.py file
+from collectorSite import app
+
+if __name__ == "__main__":
+    app.run(debug=True)
